pyproject.toml
[project]
name = "agent-reasoning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "google-genai>=0.6.0",
    "ipykernel>=6.29.5",
    "ollama>=0.4.6",
    "openai>=1.59.9",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.2",
]

[project.scripts]
evaluate = "agent_reasoning.cli.evaluate:main"
clone-repos = "agent_reasoning.cli.clone_repos:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_reasoning*"]
